more trace widths
more autorouter optimization

git-svn-id: http://fritzing.googlecode.com/svn/trunk@4619 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/items/wire.cpp
@@ -53,8 +53,8 @@ QHash<QString, QString> Wire::colors;
 QHash<QString, QString> Wire::shadowColors;
 QHash<QString, QString> Wire::colorTrans;
 QStringList Wire::colorNames;
-QHash<long, QString> Wire::widthTrans;
-QList<long> Wire::widths;
+QHash<int, QString> Wire::widthTrans;
+QList<int> Wire::widths;
 qreal Wire::STANDARD_TRACE_WIDTH;
 qreal Wire::HALF_STANDARD_TRACE_WIDTH;
